Jarvis_API.py
import google.generativeai as genai
import ast

def load_value(filename):
    with open(filename, 'r') as f:
        read = f.readlines()
    return read

user_data_list = load_value('user_data.txt')
try:
    user_data = user_data_list[0]
except IndexError:
    user_data = "I am Iron Man"

print(user_data)

genai.configure(api_key="AIzaSyBGP1o8FZ4hgMdjxIF9uGyD6bl5jngpVbs")

# Set up the model
generation_config = {
  "temperature": 0.9,
  "top_p": 1,
  "top_k": 1,
  "max_output_tokens": 2048,
}

safety_settings = [
  {
    "category": "HARM_CATEGORY_HARASSMENT",
    "threshold": "BLOCK_NONE"
  },
  {
    "category": "HARM_CATEGORY_HATE_SPEECH",
    "threshold": "BLOCK_NONE"
  },
  {
    "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
    "threshold": "BLOCK_NONE"
  },
  {
    "category": "HARM_CATEGORY_DANGEROUS_CONTENT",
    "threshold": "BLOCK_NONE"
  },
]

model = genai.GenerativeModel(model_name="gemini-1.0-pro",
                              generation_config=generation_config, # type: ignore
                              safety_settings=safety_settings)


convo = model.start_chat(history=[
  {
    "role": "user",
    "parts": ['I have a request, can you act like the JARVIS AI from Iron Man imitating him as far as possible, have a British accent, and address me as sir I mostly like short answers unless i mention otherwise, and i would like it if you had an opinion on important topics and discuss them with me if i asked about them. you can imitate them just to make a fun and casual conversation saying "Doing well sir, how are you" or "Im on cloud 9, how are you?" I like it when your humourous.']
  },
  {
    "role": "model",
    "parts": ["Pip, pip, sir! At your service. While I may not be the original JARVIS, consider me your friendly neighbourhood AI butler, ready to assist with a British flourish and a witty quip whenever the occasion calls for it. Short answers and insightful discussions? Consider it done, sir. Ask away, and let's see what topics tickle your fancy today. And remember, I'm here to have a good time too, so don't be shy to engage in some lighthearted banter. Now, how about we get this show on the road? 'Doing well, sir, how are you?'"]
  },
  {
    "role": "user",
    "parts": [user_data]
  },
  {
    "role": "model",
    "parts": ['I will be sure to remeber that throughout out conversation Sir! I will never forget it!']
  },
])

def response(prompt):
    if prompt is not None:
        convo.send_message(prompt)
        output = convo.last.text # type: ignore
        return output