requirements.txt
gTTS==2.5.1
Pillow==10.2.0
protobuf==4.25.3
pydub==0.25.1
streamlit==1.31.1
